tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/delete.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.delete tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#delete-a-stylepropertymap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<body>
<script>
'use strict';

test(t => {
  let styleMap = createInlineStyleMap(t, '');
  assert_throws(new TypeError(), () => styleMap.delete('lemon'));
  assert_throws(new TypeError(), () => styleMap.delete(null));
}, 'Calling StylePropertyMap.delete with an unsupported property name throws a TypeError');

test(t => {
  const styleMap = createInlineStyleMap(t, '');
  assert_false(styleMap.has('width'));
  styleMap.delete('width');
  assert_false(styleMap.has('width'));
}, 'Deleting a property not in a StylePropertyMap is a no-op');

const gTestCases = [
  { style: 'width: 10px', property: 'width', desc: 'a property' },
  { style: '--foo: auto', property: '--foo', desc: 'a custom property' },
  { style: 'transition-duration: 1s, 2s, 3s', property: 'transition-duration', desc: 'a list-valued property' },
  { style: 'width: 10px', property: 'wIdTh', desc: 'a mixed-case property' },
];

for (const {style, property, desc} of gTestCases) {
  test(t => {
    const styleMap = createInlineStyleMap(t, style);
    assert_true(styleMap.has(property));
    styleMap.delete(property);
    assert_false(styleMap.has(property));
  }, 'Deleting ' + desc + ' in the property model removes it from the property model');
}

</script>
